[repos] Move convenience functions to repoutils.go
