+++ core/repos/repoutils.go
@@ -1,0 +1,45 @@
+/*
+ * === This file is part of ALICE O² ===
+ *
+ * Copyright 2021 CERN and copyright holders of ALICE O².
+ * Author: Kostas Alexopoulos <kostas.alexopoulos@cern.ch>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * In applying this license CERN does not waive the privileges and
+ * immunities granted to it by virtue of its status as an
+ * Intergovernmental Organization or submit itself to any jurisdiction.
+ */
+
+package repos
+
+import (
+	"gopkg.in/yaml.v3"
+	"io/ioutil"
+)
+
+func IsFilePublicWorkflow(filePath string) bool {
+	yamlFile, err := ioutil.ReadFile(filePath)
+	if err != nil { return false }
+	return IsPublicWorkflow(yamlFile)
+}
+
+func IsPublicWorkflow(yamlFile []byte) bool {
+	var nameNode struct {
+		Node yaml.Node `yaml:"name"`
+	}
+	err := yaml.Unmarshal(yamlFile, &nameNode)
+	if err != nil { return false }
+	return nameNode.Node.Tag == "!public"
+}
